more refactorings of the external library manager
